Merge #55, Enhance #94
Explicit is better than implicit. (From the Zen of Python)
Thus I think it's better to point out that you want to match all URL
starting with a prefix into a model class.

So, the example in #94 would be refined into:
```swift
override class func modelClassesByResourcePath() -> [NSObject: AnyObject] {
    let modelClasses: [NSObject: AnyObject] = [
        "inventory/**": Item.self,
    ]
    return modelClasses
}
```
(Add a `**` to match any further paths, which is a feature proposed in #55)